create separate folders for unit/system tests
